backend/requirements.txt
Django==5.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
django-environ==0.11.2
psycopg[binary]==3.2.3
gunicorn==23.0.0
requests==2.32.3
